+++ tests/integration/lux-user-timing.spec.ts
@@ -1,4 +1,4 @@
-import { getElapsedMs, parseNestedPairs } from "../helpers/lux";
+import { getElapsedMs, parseUserTiming } from "../helpers/lux";
 
 describe("LUX user timing", () => {
   const luxRequests = requestInterceptor.createRequestMatcher("/beacon/");
@@ -10,12 +10,13 @@ describe("LUX user timing", () => {
   test("user timing marks and measures are collected in auto mode", async () => {
     await navigateTo("/user-timing.html");
     const beacon = luxRequests.getUrl(0);
-    const UT = parseNestedPairs(beacon.searchParams.get("UT"));
+    const UT = parseUserTiming(beacon.searchParams.get("UT"));
 
     expect(Object.values(UT).length).toEqual(3);
-    expect(parseInt(UT["first-mark"])).toBeGreaterThan(0);
-    expect(parseInt(UT["test-mark"])).toBeGreaterThanOrEqual(parseInt(UT["first-mark"]));
-    expect(parseInt(UT["test-measure"])).toBeGreaterThan(0);
+    expect(UT["first-mark"].startTime).toBeGreaterThan(0);
+    expect(UT["test-mark"].startTime).toBeGreaterThanOrEqual(UT["first-mark"].startTime);
+    expect(UT["test-measure"].startTime).toEqual(UT["first-mark"].startTime);
+    expect(UT["test-measure"].duration).toBeGreaterThan(0);
   });
 
   test("user timing marks and measures are collected in a SPA", async () => {
@@ -28,15 +29,32 @@ describe("LUX user timing", () => {
     await page.evaluate("LUX.send()");
 
     const beacon = luxRequests.getUrl(0);
-    const userTiming = parseNestedPairs(beacon.searchParams.get("UT"));
+    const UT = parseUserTiming(beacon.searchParams.get("UT"));
 
     // The mark and measure values will vary from test to test, so there is ~10ms margin of error.
     // To test the mark, we get the current timestamp just before creating the mark. To test the
     // measure, we wait 30ms between the mark and measure, which should make the measure ~30ms.
-    expect(parseInt(userTiming["test-mark"])).toBeGreaterThanOrEqual(timeBeforeMark);
-    expect(parseInt(userTiming["test-mark"])).toBeLessThan(timeBeforeMark + 10);
-    expect(parseInt(userTiming["test-measure"])).toBeGreaterThanOrEqual(30);
-    expect(parseInt(userTiming["test-measure"])).toBeLessThan(40);
+    expect(UT["test-mark"].startTime).toBeGreaterThanOrEqual(timeBeforeMark);
+    expect(UT["test-mark"].startTime).toBeLessThan(timeBeforeMark + 10);
+    expect(UT["test-measure"].startTime).toEqual(UT["test-mark"].startTime);
+    expect(UT["test-measure"].duration).toBeGreaterThanOrEqual(30);
+    expect(UT["test-measure"].duration).toBeLessThan(40);
+  });
+
+  test("the most recent mark takes priority over previous marks with the same name", async () => {
+    await navigateTo("/default.html?injectScript=LUX.auto=false;");
+
+    await page.evaluate("performance.mark('test-mark')");
+    await page.waitForTimeout(30);
+    const timeBeforeMark = await getElapsedMs(page);
+    await page.evaluate("performance.mark('test-mark')");
+    await page.evaluate("LUX.send()");
+
+    const beacon = luxRequests.getUrl(0);
+    const UT = parseUserTiming(beacon.searchParams.get("UT"));
+
+    expect(UT["test-mark"].startTime).toBeGreaterThanOrEqual(timeBeforeMark);
+    expect(UT["test-mark"].startTime).toBeLessThan(timeBeforeMark + 10);
   });
 
   test("user timing marks in a SPA are relative to the previous LUX.init call", async () => {
@@ -49,10 +67,10 @@ describe("LUX user timing", () => {
     await page.evaluate("LUX.send()");
 
     const beacon = luxRequests.getUrl(1);
-    const userTiming = parseNestedPairs(beacon.searchParams.get("UT"));
+    const UT = parseUserTiming(beacon.searchParams.get("UT"));
 
-    expect(parseInt(userTiming["test-mark"])).toBeGreaterThanOrEqual(30);
-    expect(parseInt(userTiming["test-mark"])).toBeLessThan(40);
+    expect(UT["test-mark"].startTime).toBeGreaterThanOrEqual(30);
+    expect(UT["test-mark"].startTime).toBeLessThan(40);
   });
 
   test("global state is not affected by LUX", async () => {
@@ -79,16 +97,16 @@ describe("LUX user timing", () => {
     await page.evaluate("performance.measure('second-test-measure', 'second-test-mark')");
     await page.evaluate("LUX.send()");
 
-    const firstUserTiming = parseNestedPairs(luxRequests.getUrl(0).searchParams.get("UT"));
-    expect(firstUserTiming).toHaveProperty("first-test-mark");
-    expect(firstUserTiming).toHaveProperty("first-test-measure");
-    expect(firstUserTiming).not.toHaveProperty("second-test-mark");
-    expect(firstUserTiming).not.toHaveProperty("second-test-measure");
+    const firstUT = parseUserTiming(luxRequests.getUrl(0).searchParams.get("UT"));
+    expect(firstUT).toHaveProperty("first-test-mark");
+    expect(firstUT).toHaveProperty("first-test-measure");
+    expect(firstUT).not.toHaveProperty("second-test-mark");
+    expect(firstUT).not.toHaveProperty("second-test-measure");
 
-    const secondUserTiming = parseNestedPairs(luxRequests.getUrl(1).searchParams.get("UT"));
-    expect(secondUserTiming).not.toHaveProperty("first-test-mark");
-    expect(secondUserTiming).not.toHaveProperty("first-test-measure");
-    expect(secondUserTiming).toHaveProperty("second-test-mark");
-    expect(secondUserTiming).toHaveProperty("second-test-measure");
+    const secondUT = parseUserTiming(luxRequests.getUrl(1).searchParams.get("UT"));
+    expect(secondUT).not.toHaveProperty("first-test-mark");
+    expect(secondUT).not.toHaveProperty("first-test-measure");
+    expect(secondUT).toHaveProperty("second-test-mark");
+    expect(secondUT).toHaveProperty("second-test-measure");
   });
 });