+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 2022-05-??: v302
 
+### Improvements
+
+- The start time of user timing measures is now recorded. In previous versions of lux.js only the duration was recorded.
 ### Bug fixes
 
 - `LUX.mark()` and `LUX.measure()` are now fully compatible with their native counterparts `performance.mark()` and `performance.measure()`. In previous versions of lux.js these functions did not support an options object as the second parameter.